refactor UrlService test
